Added UserAnswer table/service for visualization of the user's answer to a question. Refactoring on services. Added Validation on Index/Dashboard to check if user time has expired

+++ ITest/ITest.Web/Areas/Admin/Models/ManageViewModels/TestResultViewModel.cs
@@ -6,7 +6,7 @@
         public string UserName { get; set; }
         public string CategoryName { get; set; }
         public int RequestedTime { get; set; }
-        public float ExecutionTime { get; set; }
+        public double ExecutionTime { get; set; }
         public bool IsPassed { get; set; }
     }
 }
